chore: fix broken send request  button bug in developer menu
